feat(apis): proof of concept pastebin interface
I've already written a dozen of interfaces for various websites but I don't
feel like most of them are as robust as I'd like them to be. I'll wait some
more time before uploading them. I'm also still working on the indexer and
gateway so they will have to be tweaked anyways.